refactor(transfer): split the 4k-line worker module into a workers/ package
`flexkv/transfer/worker.py` had grown to ~4060 lines holding every worker class
plus the handle, the runtime loop and the dispatch registry. It is split along
the boundaries that already existed inside it:

  workers/gpu_cpu.py   GPU<->CPU transfers (both pools)
  workers/cpu_ssd.py   CPU<->SSD
  workers/gds.py       GPU<->SSD via GDS
  workers/peer.py      peer-to-peer
  workers/remote.py    mooncake-store
  workers/handle.py    WorkerHandle: the parent-side process handle
  workers/runtime.py   the child-side loop and dispatch registry

Dispatch is now a registry keyed by (PoolId, TransferType) rather than a chain
of is_swa branches, which is what makes one worker able to serve both pools.

Moves, not rewrites: the behaviour lives in the same functions, and worker.py
stays as the import surface so existing `from flexkv.transfer.worker import X`
keeps resolving.
